requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic[email]==2.5.2
python-dotenv==1.0.0
alembic==1.12.1
opencv-python==4.8.1.78
numpy==1.26.2
pandas==2.1.3
python-dateutil==2.8.2
pytest==7.4.3,
httpx==0.25.2
pillow==10.1.0
qrcode==7.4.2
reportlab==4.0.7